AIPong/Algorithm_1/requirements.txt
tensorflow-gpu==1.13.0
opencv-python
numpy
pygame